+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-### NEXT_VERSION_TYPE=MINOR
+### NEXT_VERSION_TYPE=MAJOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+**breaking changes**
+Changed semantic of `QueueService#start`. Now it must be called once at start of application to initialize queues.
+To unpause task processing use `QueueService#unpause` instead of `QueueService#start`.
+**new features**
 * Added method `QueueService#resizeQueueExecutionPool` to dynamically allocate or
 dispose threads for particular queue.
+* Added method `QueueService#unpause` and `QueueService#unpause(QueueId)` to continue tasks processing which
+have been paused by `QueueService#pause` method.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [13.1.0]() (26-08-2021)